requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
djangorestframework==3.14.0
Pillow==10.0.1
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
